requirements.txt
Flask==2.3.3
python-telegram-bot==13.15
solana==0.36.7
solders==0.26.0
schedule==1.2.0
requests==2.31.0
python-dotenv==1.0.0
beautifulsoup4==4.12.2
websocket-client==1.6.4
gunicorn==21.2.0
Werkzeug==2.3.7
